+++ tests/unit/datarobot_drum/conftest.py
@@ -246,94 +246,6 @@ def df_to_temporary_file() -> callable:
 # DRUM HELPERS FUNCS
 
 
-@pytest.fixture
-def this_dir():
-    return str(Path(__file__).absolute().parent)
-
-
-@pytest.fixture
-def target():
-    return "some-target"
-
-
-@pytest.fixture
-def model_metadata_file_factory():
-    with TemporaryDirectory(suffix="code-dir") as temp_dirname:
-
-        def _inner(input_dict):
-            file_path = Path(temp_dirname) / MODEL_CONFIG_FILENAME
-            with file_path.open("w") as fp:
-                yaml.dump(input_dict, fp)
-            return temp_dirname
-
-        yield _inner
-
-
-@pytest.fixture
-def temp_metadata(environment_id, model_metadata_file_factory):
-    metadata = {
-        "name": "joe",
-        "type": "training",
-        "targetType": "regression",
-        "environmentID": environment_id,
-        "validation": {"input": "hello"},
-    }
-    yield model_metadata_file_factory(metadata)
-
-
-@pytest.fixture
-def output_dir():
-    with TemporaryDirectory(suffix="output-dir") as dir_name:
-        yield dir_name
-
-
-@pytest.fixture
-def fit_args(temp_metadata, target, output_dir):
-    return [
-        "fit",
-        "--code-dir",
-        temp_metadata,
-        "--input",
-        __file__,
-        "--target",
-        target,
-        "--target-type",
-        "regression",
-        "--output",
-        output_dir,
-    ]
-
-
-@pytest.fixture
-def score_args(this_dir):
-    return [
-        "score",
-        "--code-dir",
-        this_dir,
-        "--input",
-        __file__,
-        "--target-type",
-        "regression",
-        "--language",
-        "python",
-    ]
-
-
-@pytest.fixture
-def server_args(this_dir):
-    return [
-        "server",
-        "--code-dir",
-        this_dir,
-        "--address",
-        "allthedice.com:1234",
-        "--target-type",
-        "regression",
-        "--language",
-        "python",
-    ]
-
-
 def set_sys_argv(cmd_line_args):
     # This is required because the sys.argv is manipulated by the 'CMRunnerArgsRegistry'
     cmd_line_args = cmd_line_args.copy()
